Split up local cluster tests into separate CI steps
