+++ ci/test-local-cluster-flakey.sh
@@ -1,0 +1,1 @@
+test-stable.sh
