+++ ci/test-local-cluster-slow.sh
@@ -1,0 +1,1 @@
+test-stable.sh
